Write an ArcGis map scraper for Parcel
Data for parcels is now scraped from the ArcGis app.
This was work was in an attempt to get the flood data mapped to a location. This effort is still not done, since the addresses aren't lat/longs yet. Although the team hasn't got the trees mapped either since there are coordinate system problems.